pyproject.toml
[project]
name = "mediaoptimizer"
version = "1.3.0.0"
description = "Media Optimizer is a tools to streamline batch  media optimization process and allow user to publish their media to cloud stroage."
authors = [
    {name = "XandeYong",email = "yongzuoxuan@gmail.com"}
]
license = {text = "GNU GENERAL PUBLIC LICENSE Version 3, 29 June 2007"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dependency-injector (==4.47.1)",
    "google-auth-oauthlib (==1.2.2)",
    "opencv-python (==4.11.0.86)",
    "pillow (==12.0.0)",
    "pillow-avif-plugin (==1.5.2)",
    "pillow-heif (==0.22.0)",
    "protobuf (==6.33.0)",
    "pydantic (==2.12.3)",
    "pytest (==8.4.2)",
    "python-magic (==0.4.27)",
    "requests (==2.32.5)",
    "tqdm (==4.67.1)",
    "python-magic-bin (==0.4.14)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
